# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Super TV API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List all packages
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
    post:
      operationId: createPackage
      tags: [packages]
      summary: Create a package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackageBody"
      responses:
        "201":
          description: Package created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
  /packages/{id}:
    put:
      operationId: updatePackage
      tags: [packages]
      summary: Update a package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackageBody"
      responses:
        "200":
          description: Updated package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
    delete:
      operationId: deletePackage
      tags: [packages]
      summary: Delete a package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Package:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        durationMinutes:
          type: integer
        price:
          type: number
        description:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, durationMinutes, price, createdAt]
    CreatePackageBody:
      type: object
      properties:
        name:
          type: string
        durationMinutes:
          type: integer
        price:
          type: number
        description:
          type: string
          nullable: true
      required: [name, durationMinutes, price]
...