# lib/api-zod/src/generated/types/package.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Super TV API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Package {
  id: number;
  name: string;
  durationMinutes: number;
  price: number;
  description?: string | null;
  createdAt: string;
}